# Console reporter: properly account for the lenght of custom counter n…